[SOIN] Extrais `<BandeauActions />`
… pour gagner en lisibilité.
