+++ svelte/lib/tableauDesMesures/TableauDesMesures.svelte
@@ -1,16 +1,14 @@
 <script lang="ts">
-  import type {
-    IdCategorie,
-    IdMesureGenerale,
-    IdService,
-    MesureGenerale,
-    MesureSpecifique,
+  import {
+    EtatEnregistrement,
+    type IdCategorie,
+    type IdMesureGenerale,
+    type IdService,
   } from './tableauDesMesures.d';
   import LigneMesure from './ligne/LigneMesure.svelte';
   import {
     enregistreMesureGenerale,
     enregistreMesuresSpecifiques,
-    metEnFormeMesures,
     recupereMesures,
   } from './tableauDesMesures.api';
   import { onMount } from 'svelte';
@@ -31,12 +29,8 @@
   import { storeNotifications } from '../ui/stores/notifications.store';
   import Avertissement from '../ui/Avertissement.svelte';
   import TagStatutMesure from '../ui/TagStatutMesure.svelte';
-
-  enum EtatEnregistrement {
-    Jamais,
-    EnCours,
-    Fait,
-  }
+  import BandeauActions from './BandeauActions.svelte';
+  import { afficheTiroirDeMesure } from './actionsTiroir';
 
   const { Jamais, EnCours, Fait } = EtatEnregistrement;
 
@@ -89,34 +83,6 @@
     ) {
       $rechercheParAvancement = 'enAction';
     }
-  };
-
-  type MesureAEditer = {
-    mesure: MesureSpecifique | MesureGenerale;
-    metadonnees: {
-      typeMesure: 'GENERALE' | 'SPECIFIQUE';
-      idMesure: string | number;
-    };
-  };
-  const afficheTiroirDeMesure = (mesureAEditer?: MesureAEditer) => {
-    document.body.dispatchEvent(
-      new CustomEvent('svelte-affiche-tiroir-ajout-mesure-specifique', {
-        detail: {
-          mesuresExistantes: metEnFormeMesures($mesures),
-          titreTiroir:
-            mesureAEditer && mesureAEditer.metadonnees.typeMesure === 'GENERALE'
-              ? mesureAEditer.mesure.description
-              : '',
-          ...(mesureAEditer && { mesureAEditer }),
-        },
-      })
-    );
-  };
-
-  const afficheTiroirExportDesMesures = () => {
-    document.body.dispatchEvent(
-      new CustomEvent('svelte-affiche-tiroir-export-mesures')
-    );
   };
 
   const marqueNouveauteLue = async () => {
@@ -213,30 +179,7 @@
       </th>
     </tr>
     {#if !estLectureSeule}
-      <tr>
-        <th colspan="2">
-          <div class="actions">
-            <button
-              class="bouton-action-mesure"
-              on:click={() => afficheTiroirDeMesure()}
-              disabled={etatEnregistrement === EnCours}
-            >
-              <img src="/statique/assets/images/icone_plus_gris.svg" alt="" />
-              Ajouter une mesure
-            </button>
-            <button
-              class="bouton-action-mesure"
-              on:click={afficheTiroirExportDesMesures}
-            >
-              <img src="/statique/assets/images/icone_export_gris.svg" alt="" />
-              Exporter la liste des mesures
-            </button>
-            {#if etatEnregistrement === EnCours}
-              <p class="enregistrement-en-cours">Enregistrement en cours ...</p>
-            {/if}
-          </div>
-        </th>
-      </tr>
+      <BandeauActions {etatEnregistrement} />
     {/if}
     <tr class="titres">
       <th>Mesure</th>
@@ -263,10 +206,7 @@
           on:click={() =>
             afficheTiroirDeMesure({
               mesure,
-              metadonnees: {
-                typeMesure: 'GENERALE',
-                idMesure: id,
-              },
+              metadonnees: { typeMesure: 'GENERALE', idMesure: id },
             })}
           estLectureSeule={estLectureSeule || etatEnregistrement === EnCours}
         />
@@ -321,55 +261,6 @@
     padding: 0;
   }
 
-  .enregistrement-en-cours {
-    font-size: 1.1em;
-    font-weight: 500;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    color: #0c5c98;
-  }
-
-  .enregistrement-en-cours:before {
-    content: '';
-    display: flex;
-    width: 24px;
-    height: 24px;
-    background-size: contain;
-    background-repeat: no-repeat;
-    background-image: url('/statique/assets/images/icone_enregistrement_en_cours.svg');
-    animation: rotation 1s linear infinite;
-  }
-
-  @keyframes rotation {
-    100% {
-      transform: rotate(360deg);
-    }
-  }
-
-  .bouton-action-mesure {
-    padding: 6px 8px;
-    display: flex;
-    align-items: center;
-    gap: 6px;
-    font-weight: 500;
-    color: var(--gris-fonce);
-    cursor: pointer;
-    background: none;
-    border: none;
-    border-radius: 4px;
-  }
-
-  .bouton-action-mesure:hover {
-    background: var(--fond-gris-pale);
-    color: var(--bleu-anssi);
-  }
-
-  .bouton-action-mesure:hover img {
-    filter: brightness(0) invert(21%) sepia(87%) saturate(646%)
-      hue-rotate(168deg) brightness(89%) contrast(103%);
-  }
-
   .tableau-des-mesures {
     border-collapse: collapse;
   }
@@ -380,22 +271,6 @@
 
   colgroup .statut-mesure {
     width: 20%;
-  }
-
-  .actions {
-    display: flex;
-    align-items: center;
-    gap: 4px;
-    padding: 18px 16px;
-    border: 1px solid var(--liseres-fonce);
-    border-radius: 0 4px 0 0;
-    transform: translateY(1px);
-    margin-left: -0.5px;
-    margin-right: -0.5px;
-  }
-
-  .actions p {
-    margin: 0;
   }
 
   thead tr th {
